+++ seapipe/pac/__init__.py
@@ -3,9 +3,8 @@ from .cfc_func import circ_wwtest, circ_kappa, mean_amp, klentropy, cohend
 from .plots import plot_mean_amps, plot_prefphase, plot_prefphase_group, plot_meanamps_group
 from .synchrony import event_sync, event_sync_dataset
 
-__all__ = ['circ_wwtest', 'circ_kappa', 'mean_amp', 'klentropy', 'cohend', 'pac_it',
-           'cfc_grouplevel', 'generate_adap_bands', 'watson_williams',
-           'generate_adap_bands', 'watson_williams',
-           'plot_mean_amps', 'plot_prefphase', 'plot_prefphase_group', 'plot_meanamps_group',
+__all__ = ['circ_wwtest', 'circ_kappa', 'mean_amp', 'klentropy', 'cohend',
+           'plot_mean_amps', 'plot_prefphase', 
+           'plot_prefphase_group', 'plot_meanamps_group',
            'event_sync', 'event_sync_dataset']