+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Beta Version
 0.3.4 More bug fixes in spindle detection for when running only some subjects at a time
 0.3.5 Added functionality to read sub and ses from tracking sheet
 0.3.7 Removed redundant scripts from pac
-0.3.9 Added first artefact detection
+0.3.9 Added first artefact detection (SAND)
 0.4 Added continuous PAC and fixed bugs with event PAC
 0.4.1 Updated format bids to allow subject-level exectution
 0.4.2 Updates to export PAC summary
@@ -22,9 +22,13 @@ Beta Version
 0.4.5 Fixed bugs with power spectrum analysis and export
 0.4.6 Fixed bugs with audit, whales, logs and added first REMS detection (YASA)
 0.4.7 Corrected some more issues with logging, updates to artefact detection
-0.4.8 Added artefact detection, fixed some loading issues, added back arousals to export macro stats
+0.4.8 Updated artefact detection, fixed some loading issues, added back arousals to export macro stats
 0.4.9 Updates to make_bids (added 'Woolcock' option and fixed bugs), whales
 0.5.0 Updates to functions (make_bids, load, audit) and enabled data_dir as /sourcedata
 0.5.1 Fixed bugs with channel renames (PSA and PAC) and make_bids
-0.5.2 Fixed bugs, added function to infer polarity, and updated other (filtering) functions
+0.5.2 Fixed bugs, added infer polarity function, and updated other (filtering) functions
 0.5.3 Fixed bugs in PAC, misc and sleep staging
+0.5.4 Fixed bugs relating to filetype in export eventparams and updated SAND
+0.6 Added S.Q.U.I.D for channel QC, and improved artefact detection
+0.6.1 Minor bug fixes, expanded default QC channel search
+0.6.2 Updates to PAC event dataset